refactor(stage-ui): abstract provider validation into a composable
Introduces the `useProviderValidation` composable to centralize the UI logic for validating provider credentials.

Previously, each provider settings page contained duplicated code for managing validation state, debouncing user input, and displaying feedback. This made the codebase difficult to maintain and prone to inconsistencies.

This commit refactors all provider settings pages to use the new composable, which encapsulates the shared logic. This change significantly simplifies the component code, improves maintainability, and ensures a consistent validation implementation across the application.
