+++ apps/stage-tamagotchi/src/pages/settings/providers/openai-compatible.vue
@@ -10,130 +10,50 @@ import {
   ProviderSettingsContainer,
   ProviderSettingsLayout,
 } from '@proj-airi/stage-ui/components'
+import { useProviderValidation } from '@proj-airi/stage-ui/composables/useProviderValidation'
 import { useProvidersStore } from '@proj-airi/stage-ui/stores/providers'
-import { useDebounceFn } from '@vueuse/core'
 import { storeToRefs } from 'pinia'
-import { computed, onMounted, ref, watch } from 'vue'
-import { useI18n } from 'vue-i18n'
-import { useRouter } from 'vue-router'
+import { computed } from 'vue'
 
-const { t } = useI18n()
-const router = useRouter()
+const providerId = 'openai-compatible'
 const providersStore = useProvidersStore()
 const { providers } = storeToRefs(providersStore) as { providers: RemovableRef<Record<string, any>> }
 
-// Get provider metadata
-const providerId = 'openai-compatible'
-const providerMetadata = computed(() => providersStore.getProviderMetadata(providerId))
-
-// Use computed properties for settings
+// Define computed properties for credentials
 const apiKey = computed({
   get: () => providers.value[providerId]?.apiKey || '',
   set: (value) => {
     if (!providers.value[providerId])
       providers.value[providerId] = {}
-
     providers.value[providerId].apiKey = value
   },
 })
 
 const baseUrl = computed({
-  get: () => providers.value[providerId]?.baseUrl || providerMetadata.value?.defaultOptions?.().baseUrl || '',
+  get: () => providers.value[providerId]?.baseUrl || '',
   set: (value) => {
     if (!providers.value[providerId])
       providers.value[providerId] = {}
-
     providers.value[providerId].baseUrl = value
   },
 })
 
-// Validation state
-const debounceTime = 500
-const isValidating = ref(0)
-const isValid = ref(false)
-const validationMessage = ref('')
-
-// Validation
-async function validateConfiguration() {
-  if (!providerMetadata.value)
-    return
-
-  isValidating.value++
-  validationMessage.value = ''
-  const startValidationTimestamp = performance.now()
-  let finalValidationMessage = ''
-
-  try {
-    const config = {
-      apiKey: apiKey.value.trim(),
-      baseUrl: baseUrl.value.trim(),
-    }
-
-    const validationResult = await providerMetadata.value.validators.validateProviderConfig(config)
-    isValid.value = validationResult.valid
-
-    if (!isValid.value)
-      finalValidationMessage = validationResult.reason
-  }
-  catch (error) {
-    isValid.value = false
-    finalValidationMessage = t('settings.dialogs.onboarding.validationError', {
-      error: error instanceof Error ? error.message : String(error),
-    })
-  }
-  finally {
-    setTimeout(() => {
-      isValidating.value--
-      validationMessage.value = finalValidationMessage
-    }, Math.max(0, debounceTime - (performance.now() - startValidationTimestamp)))
-  }
-}
-
-const debouncedValidateConfiguration = useDebounceFn(() => {
-  if (!apiKey.value.trim()) {
-    isValid.value = false
-    validationMessage.value = ''
-    isValidating.value = 0
-    return
-  }
-  validateConfiguration()
-}, debounceTime)
-
-onMounted(() => {
-  providersStore.initializeProvider(providerId)
-
-  // Initialize refs with current values
-  apiKey.value = providers.value[providerId]?.apiKey || ''
-  baseUrl.value = providers.value[providerId]?.baseUrl || providerMetadata.value?.defaultOptions?.().baseUrl || ''
-
-  if (apiKey.value.trim())
-    validateConfiguration()
-})
-
-// Watch settings and update the provider configuration
-watch([apiKey, baseUrl], () => {
-  providers.value[providerId] = {
-    ...providers.value[providerId],
-    apiKey: apiKey.value,
-    baseUrl: baseUrl.value || providerMetadata.value?.defaultOptions?.().baseUrl || '',
-  }
-  debouncedValidateConfiguration()
-}, { deep: true })
-
-function handleResetSettings() {
-  providers.value[providerId] = {
-    ...(providerMetadata.value?.defaultOptions as any),
-  }
-  isValid.value = false
-  validationMessage.value = ''
-  isValidating.value = 0
-}
+// Use the composable to get validation logic and state
+const {
+  t,
+  router,
+  providerMetadata,
+  isValidating,
+  isValid,
+  validationMessage,
+  handleResetSettings,
+} = useProviderValidation(providerId)
 </script>
 
 <template>
   <ProviderSettingsLayout
-    :provider-name="providerMetadata?.localizedName || 'OpenAI Compatible'"
-    :provider-icon="providerMetadata?.icon"
+    :provider-name="providerMetadata?.localizedName"
+    :provider-icon-color="providerMetadata?.iconColor"
     :on-back="() => router.back()"
   >
     <ProviderSettingsContainer>
@@ -152,7 +72,7 @@ function handleResetSettings() {
       <ProviderAdvancedSettings :title="t('settings.pages.providers.common.section.advanced.title')">
         <ProviderBaseUrlInput
           v-model="baseUrl"
-          :placeholder="providerMetadata?.defaultOptions?.().baseUrl as string || 'https://api.example.com/v1/'"
+          placeholder="https://api.openai.com/v1/"
         />
       </ProviderAdvancedSettings>
 
@@ -177,8 +97,8 @@ function handleResetSettings() {
 </template>
 
 <route lang="yaml">
-  meta:
-    layout: settings
-    stageTransition:
-      name: slide
-  </route>
+meta:
+  layout: settings
+  stageTransition:
+    name: slide
+</route>
