+++ apps/stage-tamagotchi/src/pages/settings/providers/together-ai.vue
@@ -1,6 +1,4 @@
 <script setup lang="ts">
-import type { RemovableRef } from '@vueuse/core'
-
 import {
   Alert,
   ProviderAdvancedSettings,
@@ -10,125 +8,21 @@ import {
   ProviderSettingsContainer,
   ProviderSettingsLayout,
 } from '@proj-airi/stage-ui/components'
-import { useProvidersStore } from '@proj-airi/stage-ui/stores/providers'
-import { useDebounceFn } from '@vueuse/core'
-import { storeToRefs } from 'pinia'
-import { computed, onMounted, ref, watch } from 'vue'
-import { useI18n } from 'vue-i18n'
-import { useRouter } from 'vue-router'
+import { useProviderValidation } from '@proj-airi/stage-ui/composables/useProviderValidation'
 
-const { t } = useI18n()
-const router = useRouter()
-const providersStore = useProvidersStore()
-const { providers } = storeToRefs(providersStore) as { providers: RemovableRef<Record<string, any>> }
-
-// Get provider metadata
 const providerId = 'together-ai'
-const providerMetadata = computed(() => providersStore.getProviderMetadata(providerId))
 
-// Use computed properties for settings
-const apiKey = computed({
-  get: () => providers.value[providerId]?.apiKey || '',
-  set: (value) => {
-    if (!providers.value[providerId])
-      providers.value[providerId] = {}
-
-    providers.value[providerId].apiKey = value
-  },
-})
-
-const baseUrl = computed({
-  get: () => providers.value[providerId]?.baseUrl || '',
-  set: (value) => {
-    if (!providers.value[providerId])
-      providers.value[providerId] = {}
-
-    providers.value[providerId].baseUrl = value
-  },
-})
-
-// Validation state
-const debounceTime = 500
-const isValidating = ref(0)
-const isValid = ref(false)
-const validationMessage = ref('')
-
-// Validation
-async function validateConfiguration() {
-  if (!providerMetadata.value)
-    return
-
-  isValidating.value++
-  validationMessage.value = ''
-  const startValidationTimestamp = performance.now()
-  let finalValidationMessage = ''
-
-  try {
-    const config = {
-      apiKey: apiKey.value.trim(),
-      baseUrl: baseUrl.value.trim(),
-    }
-
-    const validationResult = await providerMetadata.value.validators.validateProviderConfig(config)
-    isValid.value = validationResult.valid
-
-    if (!isValid.value)
-      finalValidationMessage = validationResult.reason
-  }
-  catch (error) {
-    isValid.value = false
-    finalValidationMessage = t('settings.dialogs.onboarding.validationError', {
-      error: error instanceof Error ? error.message : String(error),
-    })
-  }
-  finally {
-    setTimeout(() => {
-      isValidating.value--
-      validationMessage.value = finalValidationMessage
-    }, Math.max(0, debounceTime - (performance.now() - startValidationTimestamp)))
-  }
-}
-
-const debouncedValidateConfiguration = useDebounceFn(() => {
-  if (!apiKey.value.trim()) {
-    isValid.value = false
-    validationMessage.value = ''
-    isValidating.value = 0
-    return
-  }
-  validateConfiguration()
-}, debounceTime)
-
-onMounted(() => {
-  // Initialize provider if it doesn't exist
-  providersStore.initializeProvider(providerId)
-
-  // Initialize refs with current values
-  apiKey.value = providers.value[providerId]?.apiKey || ''
-  baseUrl.value = providers.value[providerId]?.baseUrl || ''
-
-  if (apiKey.value.trim())
-    validateConfiguration()
-})
-
-// Watch settings and update the provider configuration
-watch([apiKey, baseUrl], () => {
-  providers.value[providerId] = {
-    ...providers.value[providerId],
-    apiKey: apiKey.value,
-    baseUrl: baseUrl.value || '',
-  }
-  debouncedValidateConfiguration()
-}, { deep: true })
-
-function handleResetSettings() {
-  providers.value[providerId] = {
-    ...(providerMetadata.value?.defaultOptions as any),
-  }
-  isValid.value = false
-  validationMessage.value = ''
-  isValidating.value = 0
-}
+const {
+  t,
+  router,
+  providerMetadata,
+  apiKey,
+  baseUrl,
+  isValidating,
+  isValid,
+  validationMessage,
+  handleResetSettings,
+} = useProviderValidation(providerId)
 </script>
 
 <template>
@@ -146,7 +40,7 @@ function handleResetSettings() {
         <ProviderApiKeyInput
           v-model="apiKey"
           :provider-name="providerMetadata?.localizedName"
-          placeholder="togetherapi-..."
+          placeholder="sk-..."
         />
       </ProviderBasicSettings>
 
@@ -178,8 +72,8 @@ function handleResetSettings() {
 </template>
 
 <route lang="yaml">
-  meta:
-    layout: settings
-    stageTransition:
-      name: slide
-  </route>
+meta:
+  layout: settings
+  stageTransition:
+    name: slide
+</route>
